requirements.txt
praw
python-dotenv 